requirements/common.txt
boxsdk==2.7.1
celery==4.4.2
beautifulsoup4==4.9.3
boto3==1.15.12
botocore==1.18.12
Django==2.2.0
djangorestframework==3.10.2
django-filter==2.0.0
googlemaps==4.4.2
psycopg2-binary==2.8.3
pdfminer==20191125
django-rest-knox==3.6.0
django-cors-headers==3.0.0
pycrypto==2.6.1
requests==2.13.0
pyOpenSSL==17.2.0
django-sslserver
redis==3.3.8
rest-social-auth == 2.2.0
mongoengine==0.10
pymongo==3.2
Pillow==8.2.0
twilio==6.45.1
zcrmsdk==2.0.6
slacker==0.13.0
django-nested-admin==3.3.0
kombu==4.6.8
django-json-widget==1.0.1
django-admin-rangefilter==0.6.0
django-reverse-admin==2.8.4
django-otp==1.0.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.9
authy==2.2.6
qrcode==6.1
pyjwt>=1.4.0
django-ckeditor==6.0.0
django-storages==1.10.1
django-multiselectfield==0.1.12
PyYAML==5.4.1
django-admin-interface==0.15.5
django-import-export==2.5.0
drf-api-logger==1.0.8
django-reversion==3.0.9
django-reversion-compare==0.14.0
# django-simple-history==2.12.0
# -e git+ssh://git@code.amazatic.net/eco-farm/eco-farm-zoho-integration.git#egg=pyzoho
-e git+ssh://git@github.com/Eco-Farm/pyzoho.git#egg=pyzoho
